concatenate hashfunctions
